+++ packages/ui/src/core/accordion-header/accordion-header.props.d.ts
@@ -3,6 +3,7 @@ import type { IconProps } from '@scaleflex/icons/icon.props';
 export interface AccordionHeaderProps extends React.HTMLAttributes<HTMLDivElement> {
   label: React.ReactNode;
   expanded?: boolean;
+  hideIcon?: boolean;
   onChange?: (value: boolean) => void;
   iconProps?: IconProps;
 }