+++ packages/ui/src/core/switcher/switcher.props.d.ts
@@ -5,7 +5,7 @@ import { SwitcherSize } from './switcher-size';
 
 export type SwitcherType = Values<typeof SwitcherSize>;
 
-export interface SwitcherProps extends React.HTMLAttributes<HTMLInputElement> {
+export interface SwitcherProps extends React.HTMLAttributes<HTMLDivElement> {
   checked?: boolean;
   onChange?: (event: ChangeEvent<HTMLInputElement>) => void;
   switcherProps?: React.InputHTMLAttributes;